{cmd,pkg/common,pkg/common/composedb}: add compose history database
Add a flat file system database to the up project to allow for storing previous copies of the modified compose.yaml file. Also included the undo command to revert the current compose.yaml file to a previous version. By default, the flat file system will store the past 5 versions of the compose file.

Change-Id: Ib1f6eb1964a248c81dc237936bc5aff330cb488e