{cmd, test}: use `docker compose` command instead of `docker-compose`
Docker has changed the command for running the latest version of
compose from `docker-compose` to `docker compose`. To take advantage
of some features that are used in storj-up, including the scale
command, the new `docker compose` command must be used. On some
operating systems, this is done for you automatically via alias, but
on others, you must explicitely use the new command. Therefore, the
code and documentation will reference the new command to be correct
for all users.

Change-Id: I37674d007118e87459069d62c269b05b93ed4124