cmd/files/docker: update base images
Updating the base and build images to use a newer version of the codebase (Stroj and Gateway-mt repo) as well as updating the version of Ubuntu to the 22.04 LTS release. This ubuntu update is needed to update the version of go used during the build of the codebase within docker.

Change-Id: I2cffdff20e3ebcd21130c862bd588d0a46009636