Dlm/mailservice args (#28)
* cmd/files/docker: update base images

Updating the base and build images to use a newer version of the codebase (Stroj and Gateway-mt repo) as well as updating the version of Ubuntu to the 22.04 LTS release. This ubuntu update is needed to update the version of go used during the build of the codebase within docker.

Change-Id: I2cffdff20e3ebcd21130c862bd588d0a46009636

* cmd/files/templates/docker-compose.template.yaml: add missing ENVs

A recent change that copied the mailservice to the satellite core from the satellite API requires environment variables to configure the service. These args were just duplicated from the satellite-api service in the compose template file.

Change-Id: I531d6ca165a47cec5f4f4740083e399ba60a8ea5