credentials: autodetect satellite-api/authservice inside the containers
`storj-up credentials` command requires to set satellite-api addres (and if you use `--s3` to get S3 credentials, the authservice URL is required, too).

Defaults are adjusted for using the command from the host.

This patch makes it easier to use storj-up from the container as it sets environment variables to provide an environment specific default.

Change-Id: I2e44e687a605035bff30a37163ef9174b6d0aa31