cmd/localbin.go: fix pathing for windows
If a windows path is provided as the local path to mount the binary or front end source directory, it will have invalid backslashes not recognized by docker, unless you escape them manually. This fix allows for using the default windows path with no escaping required, however the path must be provided in quotes.

Change-Id: I0ddde0c5a6d44976cf074455c4aa56c21ed35b69