cmd/local-{entrypoint,bin}: don't create host path for bind mounts
In these commands, the CreateBind helper is being called to create a
bind mount to files, not a directory. The helper sets the
create_host_path option to true which means that docker will end up
creating the source path on the host if it does not exist.

There are two problems with this:
1. The host path that it creates is a directory and not a file. This
   leads to weird debugging messages when the containers are started
   since they bind the newly created, empty directory, which is supposed
   to be a file (e.g., bin/satellite)
2. The directories created on the host are owned by root. We shouldn't
   be polluting what is likely a user-owned directory with something
   owned by root.

This change updates CreateBind to no longer set create_host_path. It
also changes the local-bin and local-endpoint commands to fail if the
source does not exist or is not a regular file.

It also fixes local-endpoint to only add the volume if it does not
already exist.

Change-Id: I34585d3bd812677fe6720da4ff169a0798858f9e