cmd/container: fix volume target path for local-bin
The `storj-up local-bin` command has been modified to use the Unix path
separator when generating volume target paths. Previously, the host
OS's path separator was used which would produce an invalid path when
running the command on a Windows machine. This would cause the volume
not to be mounted properly.