cmd/{factory,init},runtime/standalone: Add storj and gateway env vars
The environment variables STORJ_PROJECT_DIR and GATEWAY_PROJECT_DIR can
now be used to independently specify the location of the repos when
using the standalone runtime. These variables replace the old
STORJUP_PROJECT_DIR environment variable. Additionally, the new
variables can be specified as flags, which will override the values of
the environment variables.

Change-Id: Ie82134a1c5223ccc71bf9a2f6443ee10cd54c443