runtime/compose/compose: fix labels for recipe modification
Currently the storj-up runtime will use the labels specified in the
recipe to determine if an existing service needs to be modified.
However the labels specified in the recipe, do not remain with the
service after initialization. This change will attempt to reapply the
label to the service on reload so that it may be used to determine if
the existing service needs to be modified by a new service recipe.

Change-Id: I8de1dfd0c7c47b4b18bf5e98a3a46a314507c29a