rename STORJ_ROLE to STORJUP_ROLE
This change is suggested by Ethan during the review of: https://review.dev.storj.io/c/storj/storj/+/10573

It's better to communicate that this environment variable is only for `storj-up` especially if we use it in production container images.

Unfortunatelly it's not fully backward compatible change.

Which this patch the old images won't be compatible with the storj-up with this patch.

But new images will be compatible with old storj-up, so probably it's not a problem if we release new images before we release next storj-up.

Change-Id: Idbe4921ef66d64a52d1163dfddd1d20e54f8513f