pkg/{recipe,files/templates}: decrease minimum storage node disk space
This change decreases the amount of disk space that a storage node must
have allocated in order to be selected by the satellite.

Previously, we didn't configure this value in our recipe and template,
so the satellite would use the default value of 500MB. However, commit
storj/storj@a224d4a increased the default from 500MB to 5GB, preventing
storj-up storage nodes (all of which are configured with 1GB) from
being selected, causing uploads to fail.

Change-Id: Ie9c999324b6eba43305b1e3105d71f65072e5c8f