scripts: Add script to bump storj/edge deps
Add script to automatize the straightforward process of bump the
storj.io/storj and storj.io/edge dependencies.

Depending of the changes, the bump may required manual intervention
because of API changes, in such case the script cannot fix them and it
will tell the user about it.

This commit updates the README to mention about this new script.

Change-Id: I85b85eb2f4f82444658c82a77cf74877d9d4831f